chore(tests): split monolithic JsonLdTest into per-feature files
Replace the 9-method root JsonLdTest.php with 6 focused JsonLd/* test
files, one per feature:

* InputDtoIriDenormalizationTest (input DTO with IRI relation)
* ContextOutputTest (output DTO @context shape, ignored properties)
* GenIdFalseTest (genId:false at resource and nested levels)
* PolymorphicResourceCollectionTest (per-item @type in collections)
* ItemUriTemplateCollectionTest (itemUriTemplate as @id, with stateOptions)
* MultiResourceContextTest (correct shortName per ApiResource variant)

Each file scopes its setUp/recreateSchema to only the entities it uses,
and replaces issue-numbered method names with descriptive ones.
